business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/rotator/Dec2Dms.java
package eu.estcube.gsconnector.commands.rotator;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class Dec2Dms implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+d ");
        messageString.append(command.getParameter("Dec Degrees"));
        messageString.append("\n");
        return messageString;
    }
}
